+++ NEWS.md
@@ -5,6 +5,7 @@
 * `get_sample_size`:
     - Add more complete documentation on what happens when "N" column is present. 
     - Add extra check to make sure `MungeSumstats` >=1.3.9 is installed. 
+    - `compute_n` can now be a numeric vector of length equal to `nrow(dat)`. 
 
 ## Bug fixes